requirements.txt
asgiref==3.11.0
dj-database-url==3.1.0
Django==6.0.1
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
gunicorn==23.0.0
numpy==2.4.0
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.11.0
